pyproject.toml
[project]
name = "beginner_website"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "unseeyou", email = "unseeyoubestmod@gmail.com"},
]
dependencies = [
    "flask>=3.0.3",
    "python-dotenv>=1.0.1",
    "waitress>=3.0.1",
    "beautifulsoup4>=4.12.3",
    "pandas>=2.2.3",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.scripts]
install = "pdm install"
run = "py src/server.py"
start = { composite = ["install", "run"] }